refactor: Extract functions from index.js
- Following functions now living in own file:
  - `addNextJoke`
  - `bindEventListenerToRefreshButton`
- static constant UPPER_BOUND now scoped within `getNextJokeIdx`
- `showCurrentJoke` renamed to `showInitialJoke` for clarity
- `jokeFetcher`'s jokeIdx parameter now takes a default value
