+++ src/js/getNextJokeIdx.js
@@ -1,4 +1,7 @@
-export const getNextJokeIdx = (currentJokeIdx, upperBound) => {
+import { jokeFetcherUpperBounds } from './jokeFetcher';
+
+export const getNextJokeIdx = async (currentJokeIdx) => {
+  const upperBound = await jokeFetcherUpperBounds();
   let nextJokeIdx = currentJokeIdx + 1;
 
   if (nextJokeIdx > upperBound) nextJokeIdx = 0;
