+++ src/js/getNextJokeIdx.spec.js
@@ -1,14 +1,21 @@
 import { getNextJokeIdx } from './getNextJokeIdx';
+import { jokeFetcherUpperBounds } from './jokeFetcher';
+
+jest.mock('./jokeFetcher', () => ({
+  ...jest.requireActual('./jokeFetcher'),
+  jokeFetcherUpperBounds: jest.fn().mockReturnValue(29),
+}));
 
 describe('getNextJokeIdx', () => {
-  it('incremements the current joke idx by 1', () => {
-    const nextJokeIdx = getNextJokeIdx(2);
+  it('incremements the current joke idx by 1', async () => {
+    const nextJokeIdx = await getNextJokeIdx(2);
 
     expect(nextJokeIdx).toBe(3);
   });
 
-  it('resets current joke idx to 0 when upper bound is reached', () => {
-    const nextJokeIdx = getNextJokeIdx(4, 4);
+  it('resets current joke idx to 0 when upper bound is reached', async () => {
+    const upperBounds = await jokeFetcherUpperBounds();
+    const nextJokeIdx = await getNextJokeIdx(upperBounds);
 
     expect(nextJokeIdx).toBe(0);
   });
