+++ src/js/index.js
@@ -1,32 +1,13 @@
 import '../css/index.css';
-import { jokeFetcher, jokeFetcherUpperBounds } from './jokeFetcher';
+import { jokeFetcher } from './jokeFetcher';
 import { renderJoke } from './renderJoke';
-import { getNextJokeIdx } from './getNextJokeIdx';
+import { bindEventListenerToRefreshButton } from './bindEventListenerToRefreshButton';
 
-const addNextJoke = async (nextJokeIdx) => {
-  const nextJoke = await jokeFetcher(nextJokeIdx);
+const showInitialJoke = async () => {
+  const initialJoke = await jokeFetcher();
 
-  renderJoke(nextJoke);
+  renderJoke(initialJoke);
+  bindEventListenerToRefreshButton();
 };
 
-const bindEventListenerToRefreshButton = async (currentJokeIdx) => {
-  const refreshButton = document.getElementById('refresh');
-  let currentJokeState = currentJokeIdx;
-  const UPPER_BOUND = await jokeFetcherUpperBounds();
-
-  refreshButton.addEventListener('click', async (e) => {
-    e.preventDefault();
-    currentJokeState = getNextJokeIdx(currentJokeState, UPPER_BOUND);
-    addNextJoke(currentJokeState);
-  });
-};
-
-const showCurrentJoke = async () => {
-  const currentJokeIdx = 0;
-  const currentJoke = await jokeFetcher(currentJokeIdx);
-
-  renderJoke(currentJoke);
-  bindEventListenerToRefreshButton(currentJokeIdx);
-};
-
-showCurrentJoke();
+showInitialJoke();
